+++ src/main/java/com/study/totee/api/persistence/ReplyRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ReplyRepository extends JpaRepository<ReplyEntity, Integer> {
+public interface ReplyRepository extends JpaRepository<ReplyEntity, Long> {
     List<ReplyEntity> findReplyEntityByComment_CommentId(Long commentId);
     ReplyEntity findByIdAndUser(Long replyId, UserEntity user);
 }
